flask-hello-world/requirements.txt
flask-hello-world
├── src
│   ├── app.py
│   ├── templates
│   │   └── index.html
│   └── static
│       └── styles.css
├── requirements.txt
├── .env
└── README.md